backend/requirements.txt
absl-py
astunparse
certifi
charset-normalizer
contourpy
cycler
flatbuffers
fonttools
gast
google-pasta
grpcio
h5py
haversine
idna
joblib==1.3.2
keras
kiwisolver
libclang
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
namex
numpy==1.23.5
opt_einsum
optree
packaging
pandas
patsy
pillow
protobuf
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
requests
rich
scikit-learn==1.2.2
scipy
seaborn
six
statsmodels
tensorboard
tensorboard-data-server
tensorflow-cpu==2.12.0
termcolor
threadpoolctl
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
xgboost
fastapi
uvicorn